fix(drivers): clear clippy lints blocking the enviro-mon board gate
The documented manual gate for pico_enviro_mon (clippy for both
--features board-pico-enviro-mon and sim,board-pico-enviro-mon with
--deny=warnings; pre-commit does not compile this board) had been
failing on HEAD after newer clippy lints landed. All fixes are
semantics-preserving:

- bme688/compensate.rs: parenthesize the gas-resistance var1/var3
  shift expressions (precedence unchanged, now explicit), drop a
  redundant u32 cast, and replace the humidity clamp if/else with
  i32::clamp.
- ltr559.rs: drop redundant i64-to-i64 casts in the lux branches.
- sensors.rs: allow(dead_code) on the generated sensor_table include
  under the sim feature — the table is board-cfg-driven but only the
  hardware sampling paths consume it; sim sensors are synthesized.

+++ picodroid-core/src/drivers/ltr559.rs
@@ -99,13 +99,13 @@ fn compute_lux_milli(ch0: u16, ch1: u16) -> u32 {
     // lux × 1000 = (a*ch0 - b*ch1) / 10
     let lux_milli = if ratio_milli < 450 {
         // 17743 * ch0 + 11059 * ch1, ÷ 10000 → lux. Multiply by 1000 → milli-lux.
-        ((17743 * ch0 + 11059 * ch1) / 10) as i64
+        (17743 * ch0 + 11059 * ch1) / 10
     } else if ratio_milli < 640 {
         // 42785 * ch0 - 19548 * ch1
-        ((42785 * ch0 - 19548 * ch1) / 10) as i64
+        (42785 * ch0 - 19548 * ch1) / 10
     } else if ratio_milli < 850 {
         // 5926 * ch0 + 1185 * ch1
-        ((5926 * ch0 + 1185 * ch1) / 10) as i64
+        (5926 * ch0 + 1185 * ch1) / 10
     } else {
         0
     };
